config added to api

+++ Storm/Storm/Config.cs
@@ -1,10 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Text;
-using System.Threading.Tasks;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using Storm.ExternalEvent;
@@ -36,34 +34,6 @@ namespace Storm
             return null;
         }
 
-        public virtual Config UpdateConfig(Config baseConfig)
-        {
-            if (JObject == null)
-                return baseConfig;
-
-            string p = baseConfig.ConfigLocation;
-            Config uConfig = (Config)Activator.CreateInstance(baseConfig.GetType());
-            uConfig = GenerateBaseConfig(uConfig);
-            uConfig.ConfigLocation = p;
-
-            foreach (var v in JObject.Properties())
-            {
-                try
-                {
-                    uConfig.GetType().GetProperty(v.Name).SetValue(uConfig, v.Value);
-                }
-                catch
-                {
-                    //Console.WriteLine("Failed to assign property [{0}] from json to property [{0}] in Config", v.Name);
-                    //No output?
-                }
-            }
-
-            WriteConfig(uConfig);
-
-            return uConfig;
-        }
-
         public virtual Config LoadConfig(Config baseConfig)
         {
             if (!File.Exists(baseConfig.ConfigLocation))
@@ -74,38 +44,51 @@ namespace Storm
             else
             {
                 string p = baseConfig.ConfigLocation;
-                JObject j = new JObject();
 
                 try
                 {
-                    j = JObject.Parse(Encoding.UTF8.GetString(File.ReadAllBytes(baseConfig.ConfigLocation)));
+                    var j = JObject.Parse(Encoding.UTF8.GetString(File.ReadAllBytes(baseConfig.ConfigLocation)));
                     baseConfig = (Config) j.ToObject(baseConfig.GetType());
+                    baseConfig.ConfigLocation = p;
+                    baseConfig.JObject = j;
+
+                    baseConfig = UpdateConfig(baseConfig);
+                    baseConfig.ConfigLocation = p;
+                    baseConfig.JObject = j;
+
+                    WriteConfig(baseConfig);
                 }
                 catch
                 {
-                    Console.WriteLine("Invalid JSON Moved: " + p);
-                    File.Move(p, Path.Combine(Path.GetDirectoryName(p), Path.GetFileNameWithoutExtension(p) + "." + Guid.NewGuid() + ".json")); //Get it out of the way for a new one
-                    var v = (Config)baseConfig.GetType().GetMethod("GenerateBaseConfig", BindingFlags.Public | BindingFlags.Instance).Invoke(baseConfig, new object[] { baseConfig });
+                    Console.WriteLine("Invalid JSON Renamed: " + p);
+                    if (File.Exists(p))
+                        File.Move(p, Path.Combine(Path.GetDirectoryName(p), Path.GetFileNameWithoutExtension(p) + "." + Guid.NewGuid() + ".json")); //Get it out of the way for a new one
+                    var v = (Config) baseConfig.GetType().GetMethod("GenerateBaseConfig", BindingFlags.Public | BindingFlags.Instance).Invoke(baseConfig, new object[] {baseConfig});
                     WriteConfig(v);
                 }
-
-                baseConfig.ConfigLocation = p;
-                baseConfig.JObject = j;
-
-                baseConfig = UpdateConfig(baseConfig);
             }
 
             return baseConfig;
         }
 
+        public virtual Config UpdateConfig(Config baseConfig)
+        {
+            return (Config)JObject.Parse(JsonConvert.SerializeObject(baseConfig, baseConfig.GetType(), Formatting.Indented, new JsonSerializerSettings())).ToObject(baseConfig.GetType());
+        }
+
         public virtual void WriteConfig(Config baseConfig)
         {
-            File.WriteAllBytes(baseConfig.ConfigLocation, Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(baseConfig, baseConfig.GetType(), Formatting.Indented, new JsonSerializerSettings())));
+            byte[] toWrite = Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(baseConfig, baseConfig.GetType(), Formatting.Indented, new JsonSerializerSettings()));
+            if (!File.Exists(baseConfig.ConfigLocation) || !File.ReadAllBytes(baseConfig.ConfigLocation).SequenceEqual(toWrite))
+                File.WriteAllBytes(baseConfig.ConfigLocation, toWrite);
+            toWrite = null;
         }
+
+
 
         public static string GetBasePath(DiskResource theMod)
         {
-            return theMod.PathOnDisk + "\\Config.json";
+            return theMod.PathOnDisk + "\\config.json";
         }
 
         public static Config Instance { get { return new Config(); } }
